feat: created initial migration
This migration is based on the assumption that we will do an extra query to get the user from the auth table.

Doing it the way of base-lining might introduce difficulties in the future when they provide a fix and there is nolonger a need for a hackish way to get auth and public database to be connected.

Note: this logic will have to be tested out first and foremost with a login sequence